Implement `IntoIter` for `HeaderMap` (#36)
* Implement `IntoIter` for `HeaderMap`

Given the multimap characteristics, the yielded item is `(Option<HeaderName>, HeaderValue)`. This is due to the fact that `IntoIter` takes ownership of all the `HeaderMap` state and yielded Iterator items cannot have a lifetime bound to the iterator.

* Add extend impl + VacantEntry::insert_entry

The extend implementation allows extending a `HeaderMap` with another `HeaderMap`. An additional API to convert a `VacantEntry` to an `OccupiedEntry` was required.