Provide Uri accessor for Authority.
Currently, there is no way to access the `Authority` component of `Uri`
at a type level. This is desireable because:

* `Authority` is backed by `Bytes` (for clone).
* `Authority` should be case-insensitive for Eq / Hash.

Unfortunately, `Uri::authority` already exists and returns a `&str`.
This is an oversight that originated from the original iteration of
`Uri` not providing typed components.

This PR adds an `authority_part` accessor to URI and deprecates
`authority` with the expectation that the next semver release of `http`
will rename `authority_part` back to `authority`.

This PR also provides an `Eq` and `Hash` implementation for `Authority`
and adds `Authority::port`.