Adding the `QueryItem` struct in path
As requested in #206, building an URI from a builder should have a
smaller granularity.

A struct `QueryItem` has been added. It holds the
elements of a query component.

A method `new` inside `PathAndQuery` has been added.
It  that requires a `path` and a list of `QueryItem` and
build a new `PathAndQuery` from them.

The URI builder could hold both a `String` and a `Vec<QueryItem>` fields
instead of one `parts` field and could provide new methods
to configure the path and the query separately.