Add http::Error::inner_ref and cause (#303)
* add Error::inner_ref() for access to inner error type

Also use `Error::inner_ref` to implement `std::error::Error::cause`
for now. When MSRV is raised to 1.30.0+, this could also be used for
`Error::source` without a breaking change.

* add Error::is helper

* rename Error::inner_ref to get_ref

* make Error::cause be the inner error's cause

None of the inner errors currently have a cause, but this is
considered (see PR discussion) the most correct implementation of
`std::error::Error::cause` and is thus future proofed.