Remove obsolete `allow(warnings)` attributes (#373)
We were for a time allowing warnings on two sections of code due to
continued use of `...` pattern syntax.  We had deferred the change to
`..=` syntax so as to support earlier Rust versions.

However, in commit `6003c8eef9fd2e75668c88288462f8ca5d07389e` we
changed over to `..=` syntax.  The `allow(warnings)` attributes and
the related comments are now obsolete, and we remove them in this
commit.