Restructure Debug impl for HeaderValue
The new code uses Iterator::try_fold() instead of external iteration to
review the bytes in the HeaderValue and separate the printing of visible
ascii bytes printing other bytes (and special casing b'"').

This new implementation may behave slightly different for a HeaderValue
of length usize::MAX, but such a HeaderValue is likely to exhaust memory
before the Debug impl comes into play (which also makes it hard to test
this hypothetical edge case).