Add TODO marker for a HeaderValue function
HeaderValue::from_maybe_shared_unchecked() is currently an unsafe
function but it could be made a safe function without intoducing a
soundness hole because it does not do anything directly that depends on
`unsafe` and it isn't protecting any invariants of HeaderValue that are
relied upon later when using `unsafe`.

from_maybe_shared_unchecked() can accept bytes as a HeaderValue that are
invalid bytes for what is described in RFC 7230 (section 3.2) but all
such bytes would still be valid to convert directly to a str since they
are all valid single-byte UTF-8.