Improve performance of PathAndQuery::from_shared
This mostly eliminates the performance regressions from the refactoring,
but does so by reintoducing the two separate loops: one for parsing the
path and one for parsing the query.

How this version differs form the orginal two-loop version is that it
eliminates the early exits from the two loops for matching the the
fragment specifier ('#'). It instead encodes this into a take_while()
combinator. This makes it more ovious that the two loops are are
checking each byte which is relied on on the "Safety" justification.