Remove unsafe from one HeaderValue function.
HeaderValue::from_maybe_shared_unchecked had been marked unsafe but this
was to indicate caller responsibility for api constraints and not caller
responsibility for avoiding undefined behaviour. This change removes the
use of unsafe for this purpose and instead uses a more explict doc
comment to draw the caller's attention to the restrictions (as well as
the continued use of the "_unchecked" suffix.

This is not a breaking change (despite affecting the public API) because
removing "unsafe" from a function still allows existing uses to compile.

This use of "unsafe" was not protecting an invariant on the internal
structre of HeaderValue that was relined on in unsafe blocks elsewhere
in the module to ensure the absence of undefined behaviour. All (other)
unsafe blocks in this module rely on local checking to ensure they avoid
undefined behaviour.