Add safety comments to header::name module
The comments describe the pre-condition and post-conditions on the
various different parts of the two parse_hdr() implementations that
combine to make the use of MaybeUninit in that function sound.

The process of assessing the soundness of the use of MaybeUninit also
included manually checking that the number of parameters to each all
eq!() invocation matches the number of bytes initilized by the
immediatly preceeding invocation of to_lower!() (which is necessary to
avoid undefined behavior). To avoid being overly repetative the general
pattern that assures soundness is documented in the comments but not
each instance of that pattern. Each instance, though, was checked.