Implement a non-failing Response Builder.
Fixes #330.  Creating a Response with the builder pattern should not
require handling errors unless opting in by using a `try_`-method
(which returns a `Result<Builder2>`, not a `Builder2` with an interal
error state).

This is really a breaking change to `response::Builder`, but rather
than suggstig a breaking change directly, I introduce an alternative
`response::Builder2`.  If this is liked, the original `Builder` can be
deprecated, and in a later release `Builder2` can be renamed back to
`Builder`.