Non-failing Uri Builder.
Fixes one part of #330.  This makes it possible to create an `Uri`
using the builder pattern without hanlding errors.

This is really a breaking change to `uri::Builder`, but rather than
suggesting a breaking change directly, I introduce an alternative
`uri::Builder2`. If this is accepted, the original Builder is
deprecated, and in a later release Builder2 can be renamed back to
Builder.

The new `Builder2` is generic on its contents, which starts out as the
empty type `()`.  The parts `Scheme`, `Authority` and `PathAndQuery`
can be added in any combination, and the `build` method is implemented
for valid combinations only.