Add author copyrights to the header map file
The HeaderMap implementation as a whole is mostly orignal to http-rs,
but the implementation was based on and derived from the indexmap crate
(formerly ordermap) from circa 2017, from which the structure and many
methods were adapted. For this reason, mention the two indexmap authors
(one of them is me) from that time.