Add `Request::parts()` and `Request::parts_mut()`
`http::Request::into_parts()` is the only way to access a request's
`Parts` structure.

It can be convenient to write utility code that takes a `&Parts` or
`&mut Parts`, i.e., so there's no need to parameterize the utility on
body type if it can't be used.

This change introduces accessors for the `Parts` structure so callers
need not Use `Request::into_parts` and `Request::from_parts` to access
the `Parts` structure.