feat: add support for HTTP QUERY method
The HTTP QUERY method is a means of making a safe, idempotent request
that contains content. The primary motivator for this case is graphql,
replacing HTTP POST for read only requests, however it is application
agnostic.

This method is still in draft but stable enough where it is unlikely to
change for this implementation.

https://httpwg.org/http-extensions/draft-ietf-httpbis-safe-method-w-body.html