feat: add support for HTTP QUERY method
The HTTP QUERY method is a means of making a safe, idempotent request
that contains content. The primary motivator for this case is graphql,
replacing HTTP POST for read only requests, however it is application
agnostic.

This method is still in draft but since 2025-01-07 it has reached consensus
so it should be stable enough to make it unlikely to change this implementation.

https://datatracker.ietf.org/doc/draft-ietf-httpbis-safe-method-w-body/